exception handler; autoload; some class files
